Cleaning up the debugging output.
This basically just trims the amount of crap the script will spit out.
It adds an 'info' level of output, and moves a lot of stuff into that
which basically just tracks the actual progress of things without saying
in detail what's going on. It also allows all output to be disabled
except for a few initial statements (i.e. ones that don't get output
every processing loop).

This makes things more practical for running as a daemon.