Add support for specifying database names.
Since the mysql protocol doesn't allow placeholders for the database
name in queries (i.e. where you go 'nova.instances') we have to munge
the query strings before passing them to mysql - this is done using the
.format() string method, since it doesn't clash with any of the
placeholder formats (%s and %(name)s). The database name needs to be
specified using the format string field name syntax, with the field name
being the default database name (i.e. '{nova}', '{keystone}', etc).

Queries that refer to source databases will always need to have the
database name munged in this way, and when a new database is added to
the set that are queried the database name needs to be added to both the
configuration file and the default set in the
reporting_pollster.common.config module.

Change-Id: I363013b4fb3edf11705174cc7635e48a6fa990e5