Update volumes query for cinder kilo schema.
In Kilo attachment information was moved out to a separate table
(volume_attachment), requiring a join to pull in the instance_uuid
information. A left outer join is used to capture /all/ the volumes, as
was the case previously, with deleted attachments being filtered out to
ensure that only current attachment information is recorded. In
addition, to protect against multiply attached volumes (of which there
are a few, due to bugs), we select only distinct volume ids /after/ all
other filtering, so that only in the case of multiply attached volumes
are we simply picking the first of a number, rather than in any case
where multiple mounts have existed historically.

This may not really be necessary longer term, since it appears no one is
using this information, but this change is the least impact option since
it avoids changing any schemas or changing the meaning of the data in
the database.

Also updates .gitreview to point at nectar/liberty branch

Change-Id: If85544580fcf89cb28540e29556ebd660b1678e9