Removed direct usage of novaclient.v2.client.Client
It is now considered unsafe to initialize this class directly,
with the documentation stating that it should be done via
novaclient.client.Client interface. This code makes that change.
This interface expects the version number to be explicitly set
as the first argument, so a retrofit to add it to the Config's
interface has been added

Change-Id: I2386b3902b1113d3879307890db5ad2d274cb3f1