Converted comment delimiters from single quotes to double quotes
My IDE (PyCharm) is having absolute conniptions at the use of double
quotes around the comment strings in the schema file.

A look at the syntax tree for the create tables command shows that
the documentation preferences the use of single quotes.
http://dev.mysql.com/doc/refman/5.7/en/create-table.html

To make it easier for me to pick up other defects from the noise I've
changed the quotes to single quotes

Change-Id: I432f9a780d9e48063f62059f412719ecdbdefab5