Corrected exit polling loop condition
There is a defect whereby if an exception is thrown before the exit
polling  loop condition is evaluated, the loop will be repeated: and
if the same exception occurs, the loop will be repeated. Again and
again...

This  is not good behaviour if you only want the loop to run once.

This commit fixes the defect.

Change-Id: Ia3ddb9dde2812e6fb36cf08f6ee9c7d24f5ca2fd