Changed to use named placeholders on update queries
Update queries required that two conversion specifiers be included:
even if only one was needed. To work around this requirement oddities
such as putting a conversion specifier in a comment were done.

Switching to using named placeholders means it is less likely that
someone at some future date will remove a needed comment. It also
has the advantage of showing what field will be used to populate the
value when the query is run.

Change-Id: I7e08f4766ac702e4cc0465a82fe8d486e93269d4