Add primary key for the role table.
This is technically meaningless, since the primary key covers all
columns in the table. However, it's necessary because using 'replace
into' syntax requires a primary key collision before the update in place
behaviour is triggered. This has resulted in the role table balooning
with masses of duplicates, which needs to be addressed asap.

Change-Id: I60fa2e9f8f66d3cf1537573276cdd0de9a3b018e