Add additional fields to the allocations table.
For more useful reporting of allocation status we want to know the start
and end dates for the allocation. We also now include the chief
investigator from the rcallocation_chiefinvestigator table, to better
handle reporting on projects that were submitted by a user logging in
with a student email.

Finally, we widen the net for statuses, adding the 'J' status
(allocation update rejected) to the query. This status appears to
indicate an allocation that is still usable, even though the requested
update was not accepted. The details of this status need clarification,
but there are a number of active projects in this state and it needs to
be reported on.

Change-Id: Iced22970b36ebc78cead56d5ed791fcc019a14b6