Add support for updating a subset of tables.
This is intended to provide a cleaner way to do repairs to the
database: rather than update /everything/, we can target specific
tables.

Also included are signficant cleanups to error handling in the
configuration processing, removing the hard-coded list of tables
(replaced by one generated through introspection), and support for
getting at the introspected list of tables.

TODO: add knowledge of the additional tables managed by entities on
top of the tables they're directly responsible for (e.g. the
historical_usage table).

Change-Id: Idcd299c9474f543884b8740b71d08595f298b09d