Update availability zone handling for Newton.
This is primarily driven by changes to the aggregate id - Newton moves
aggregates to the top level, and the id becomes a globally unique
integer. This code makes a number of adjustments to the aggregate and
hypervisor processing that flow from that change.

This change set also includes a change to the database schema, adding a
new column to the hypervisor table.

Change-Id: If4faaf7a397fc15cfc3b9d4f9ce4735aae5e72ec