Update handling of dependent table updates.
This patch completely changes the way that handling of dependent table
updates is done - i.e. those cases where the data in one table is
derived from data in another table. At the moment this is limited to the
hypervisor availability zone (determined by the host aggregate
membership data), and the project table's 'has_instances' column which
is determined by the contents of the instance table. Previously the
updates for these tables were driven by the source of the derived data
(i.e. the aggregate_host table update also updated the hypervisor table,
and the instance table update also updated the project table). This has
been changed so that the aggregate_host update caches the data necessary
for the hypervisor table update process to insert the correct data in
the first place, and likewise the instance update for the project table.
This affects the order in which updates have to happen, basically
reversing the old order dependencies.

I've come across some correctness issues related to the project table
(the has_instances column is hard to keep properly up to date given the
nature of the update process for the project table), but it's probably
not enormously important (this obviously needs investigating).

The patch is quite large, but it's almost more comment than code.

Change-Id: Ib29f3ccd3d10a4033ac3583bc86286bb2a963349