Ensure that the instance table AZ field is correct.
This change makes use of cached data from the aggregate table update to
set the instance table availability_zone value, based on the instance
hypervisor. This is the correct method for determining the instance AZ
in more recent OpenStack. This results in a dependency chain, where
updating the project table depends on the has_instances data from an
instance table update, which in turn depends on the aggregate table
update for the hypervisor AZ data. Hence, the change is significantly
larger than the headline would suggest.

In addition to the headline change (and representing the bulk of the
code, unfortunately) this change includes support for determining the
tables to be updated and the order in which they're updated based on
dependency lists provided by each Entity derived class. Any Entity
derived class that depends on cached data from another table update
process can declare that dependency in the _get_dependencies() class
method, returning a set that contains the prerequisite tables. The list
of requested tables (the full list of supported tables by default, or a
user-specified list if one is supplied) is augmented to ensure that all
dependencies are satisfied, and then topologically sorted to allow
processing in an order that satisfies the declared dependencies. The
algorithm used is a minor variation on Kahn's algorithm, and is
documented in the code. Additional tests have been added to ensure that
the dependency processing is correct.

The decision to add dependency handling was driven by the increasing
complexity of table interrelationships and the fact that chained
dependencies had appeared. It should enable more complex processing of
the source data without requiring more and more special case orderings,
particularly in the case of updating a user-specified subset of tables.

Change-Id: I6454aa7e83bfefa5aa70b7c436c668d4cbac5791